+++ src/librpbase/disc/CBCReader.cpp
@@ -281,8 +281,8 @@ size_t CBCReader::read(void *ptr, size_t size)
 		m_lastError = EIO;
 		return 0;
 	}
-	unsigned int sz_dec = d->cipher->decrypt(
-		static_cast<uint8_t*>(ptr), (unsigned int)size);
+	size_t sz_dec = d->cipher->decrypt(
+		static_cast<uint8_t*>(ptr), size);
 	if (sz_dec != size) {
 		// decrypt() failed.
 		m_lastError = EIO;
